Added authorization for adding games. Cleaned up MyGames section.

+++ FrameAdvance/Controllers/GameController.cs
@@ -17,6 +17,7 @@ namespace FrameAdvance.Controllers
         public GameController(ApplicationDbContext context)
         {
             _gameRepository = new GameRepository(context);
+            _userProfileRepository = new UserProfileRepository(context);
         }
 
         [HttpGet]
@@ -34,6 +35,13 @@ namespace FrameAdvance.Controllers
         [HttpPost]
         public IActionResult Post(Game game)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+
+            if (currentUserProfile.UserTypeId != 1)
+            {
+                return Unauthorized();
+            }
+
             _gameRepository.Add(game);
             return Ok();
         }
@@ -41,6 +49,13 @@ namespace FrameAdvance.Controllers
         [HttpPut("{id}")]
         public IActionResult Put(int id, Game game)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+
+            if (currentUserProfile.UserTypeId != 1)
+            {
+                return Unauthorized();
+            }
+
             if (id != game.Id)
             {
                 return BadRequest();
@@ -53,6 +68,13 @@ namespace FrameAdvance.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+
+            if (currentUserProfile.UserTypeId != 1)
+            {
+                return Unauthorized();
+            }
+
             _gameRepository.Delete(id);
             return NoContent();
         }
@@ -61,6 +83,13 @@ namespace FrameAdvance.Controllers
         [HttpPost("addgame")]
         public IActionResult Post(UserGame userGame)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+
+            if (currentUserProfile.Id != userGame.UserProfileId)
+            {
+                return Unauthorized();
+            }
+
             _gameRepository.AddUserGame(userGame);
             return CreatedAtAction("Get", new { id = userGame.Id }, userGame);
         }
@@ -68,22 +97,42 @@ namespace FrameAdvance.Controllers
         [HttpPut("editusergame/{id}")]
         public IActionResult Put(int id, UserGame userGame)
         {
+            var currentUserProfile = GetCurrentUserProfile();
+
+            if (currentUserProfile.Id != userGame.UserProfileId)
+            {
+                return Unauthorized();
+            }
             if (id != userGame.Id)
             {
                 return BadRequest();
             }
+
 
             _gameRepository.UpdateUserGame(userGame);
             return NoContent();
         }
 
         [HttpDelete("removegame/{id}")]
-        public IActionResult DeletePostTag(int id)
+        public IActionResult DeleteUserGame(int id)
         {
+
+            var currentUserProfile = GetCurrentUserProfile();
+            var userGame = _gameRepository.GetUserGameById(id);
+
+            if (currentUserProfile.Id != userGame.UserProfileId)
+            {
+                return Unauthorized();
+            }
             _gameRepository.RemoveUserGame(id);
             return NoContent();
         }
 
+        private UserProfile GetCurrentUserProfile()
+        {
+            var firebaseUserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            return _userProfileRepository.GetByFirebaseUserId(firebaseUserId);
+        }
 
     }
 }
